// controls/pdfexport/src/implementation/graphics/brushes/enum.ts
export enum PdfExtend {
    /**
     * Do not extend any point.
     */
    None = 0,
    /**
     * Extend start point.
     */
    Start = 1,
    /**
     * Extend end point.
     */
    End = 2,
    /**
     * Extend both start and end points.
     */
    Both = Start | End,
}
/**
 * Specifies the gradient direction of the linear gradient brush.
 */
export enum PdfLinearGradientMode {
    /**
     * Specifies a gradient from upper right to lower left.
     */
    BackwardDiagonal,
    /**
     * Specifies a gradient from upper left to lower right.
     */
    ForwardDiagonal,
    /**
     * Specifies a gradient from left to right.
     */
    Horizontal,
    /**
     * Specifies a gradient from top to bottom.
     */
    Vertical,
}
/**
 * Shading type constants.
 */
export enum ShadingType {
    /**
     * Function-based shading.
     */
    Function = 1,
    /**
     * Axial shading.
     */
    Axial = 2,
    /**
     * Radial shading.
     */
    Radial = 3
}